# openapi/items-service.openapi.yaml
openapi: 3.0.3
info:
  title: Showcase API
  description: Minimal API showcasing custom validation with OpenAPI code generation
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: items

paths:
  /api/items:
    get:
      operationId: getItems
      tags:
        - items
      parameters:
        - name: email
          in: query
          description: Filter by email
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
    post:
      operationId: createItem
      tags:
        - items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  /api/items/{id}:
    patch:
      operationId: validatedUpdateItem
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the item
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatedUpdateItemRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
    put:
      operationId: updateItem
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the item
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

components:
  schemas:
    ItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string

    CreateItemRequest:
      type: object
      required:
        - email
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 30
        email:
          type: string
          minLength: 1
          maxLength: 30

    UpdateNameRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 30

    UpdateItemRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string

    ValidatedUpdateItemRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string

    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'

    ValidationError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string